+++ packages/swapkit/core/src/index.ts
@@ -2,4 +2,4 @@ export * from "@lastnetwork/api";
 export * from "@lastnetwork/helpers";
 export { stripToCashAddress } from "@lastnetwork/toolbox-utxo";
 
-export * from "./client.ts";
+export * from "./client";