+++ packages/toolboxes/cosmos/src/toolbox/BaseCosmosToolbox.ts
@@ -1,12 +1,18 @@
 import { Bip39, EnglishMnemonic, Slip10, Slip10Curve, stringToPath } from "@cosmjs/crypto";
 import { DirectSecp256k1HdWallet, DirectSecp256k1Wallet } from "@cosmjs/proto-signing";
 import { SwapKitApi } from "@lastnetwork/api";
-import { AssetValue, Chain, ChainId, type DerivationPath } from "@lastnetwork/helpers";
+import {
+  AssetValue,
+  Chain,
+  ChainId,
+  ChainIdToChain,
+  type DerivationPath,
+} from "@lastnetwork/helpers";
 
-import { CosmosClient } from "../cosmosClient.ts";
-import type { ToolboxParams } from "../index.ts";
-import type { BaseCosmosToolboxType } from "../thorchainUtils/types/client-types.ts";
-import { USK_KUJIRA_FACTORY_DENOM } from "../util.ts";
+import { CosmosClient } from "../cosmosClient";
+import type { ToolboxParams } from "../index";
+import type { BaseCosmosToolboxType } from "../thorchainUtils/types/client-types";
+import { USK_KUJIRA_FACTORY_DENOM } from "../util";
 
 type Params = {
   client: CosmosClient;
@@ -87,10 +93,17 @@ export const BaseCosmosToolbox = ({
   getFeeRateFromThorswap,
   getBalance: async (address: string, _potentialScamFilter?: boolean) => {
     const denomBalances = await cosmosClient.getBalance(address);
+    const chain = ChainIdToChain[cosmosClient.chainId];
     return await Promise.all(
       denomBalances
         .filter(({ denom }) => denom && !denom.includes("IBC/"))
-        .map(({ denom, amount }) => getAssetFromDenom(denom, amount)),
+        .map(({ denom, amount }) => {
+          const fullDenom =
+            [Chain.THORChain, Chain.Maya].includes(chain) && denom.includes("/")
+              ? `${chain}.${denom}`
+              : denom;
+          return getAssetFromDenom(fullDenom, amount);
+        }),
     );
   },
 });
@@ -145,7 +158,6 @@ export const cosmosValidateAddress = ({
       return client.checkAddress(address);
     }
   }
-  return false;
 };
 
 export const estimateTransactionFee = ({ assetValue }: { assetValue: AssetValue }) => {