+++ packages/toolboxes/radix/src/types/index.ts
@@ -1,9 +1,14 @@
-import { NetworkId } from "@radixdlt/radix-engine-toolkit";
+import { Chain } from "@lastnetwork/helpers";
+import type { RadixToolbox } from "../index";
 
 export const RadixMainnet = {
-  networkId: NetworkId.Mainnet,
+  networkId: 1,
   networkName: "mainnet",
   dashboardBase: "https://dashboard.radixdlt.com",
 };
 
 export type RadixNetwork = typeof RadixMainnet;
+
+export type RadixWallets = {
+  [Chain.Radix]: Awaited<ReturnType<typeof RadixToolbox>>;
+};