+++ packages/wallets/phantom/src/types.ts
@@ -1,9 +1,0 @@
-import type { SolanaProvider } from "@lastnetwork/toolbox-solana";
-
-declare global {
-  interface Window {
-    phantom: {
-      solana: SolanaProvider;
-    };
-  }
-}