+++ packages/wallets/trezor/src/index.ts
@@ -35,7 +35,7 @@ type TrezorOptions = {
 
 type Params = TrezorOptions & {
   // TODO improve api typing
-  api?: Todo;
+  api?: any;
   chain: Chain;
   derivationPath: DerivationPathArray;
   rpcUrl?: string;
@@ -72,7 +72,7 @@ async function getToolbox({
     case Chain.Sepolia:
     case Chain.Ethereum: {
       const { getProvider, getToolboxByChain } = await import("@lastnetwork/toolbox-evm");
-      const { getEVMSigner } = await import("./evmSigner.ts");
+      const { getEVMSigner } = await import("./evmSigner");
 
       const keys = ensureEVMApiKeys({ chain, ethplorerApiKey, covalentApiKey });
       const provider = getProvider(chain, rpcUrl);
@@ -157,7 +157,7 @@ async function getToolbox({
 
           // Lint is not happy with the type of txOutputs
           // biome-ignore lint/complexity/noExcessiveCognitiveComplexity: TODO: refactor
-          outputs: psbt.txOutputs.map((output: Todo) => {
+          outputs: psbt.txOutputs.map((output: any) => {
             const outputAddress =
               chain === Chain.BitcoinCash && output.address
                 ? toCashAddress(output.address)