+++ packages/wallets/xdefi/src/walletHelpers.ts
@@ -2,7 +2,7 @@ import type { Keplr } from "@keplr-wallet/types";
 import {
   type AssetValue,
   Chain,
-  type ChainId,
+  ChainId,
   ChainToChainId,
   type EVMChain,
   EVMChains,
@@ -165,7 +165,7 @@ export async function getXDEFIAddress(chain: Chain) {
     eipProvider.request(
       { method: "request_accounts", params: [] },
       // @ts-expect-error
-      (error: Todo, [response]: string[]) => (error ? reject(error) : resolve(response)),
+      (error: any, [response]: string[]) => (error ? reject(error) : resolve(response)),
     ),
   );
 }
@@ -214,9 +214,14 @@ export function cosmosTransfer({
 }) {
   return async ({ from, recipient, assetValue, memo }: TransferParams) => {
     const { createSigningStargateClient } = await import("@lastnetwork/toolbox-cosmos");
+    await window.xfi?.keplr?.enable(chainId);
     // @ts-ignore
     const offlineSigner = window.xfi?.keplr?.getOfflineSignerOnlyAmino(chainId);
-    const cosmJS = await createSigningStargateClient(rpcUrl || RPCUrl.Cosmos, offlineSigner);
+    const cosmJS = await createSigningStargateClient(
+      rpcUrl || RPCUrl.Cosmos,
+      offlineSigner,
+      chainId === ChainId.Kujira ? "0.0003ukuji" : undefined,
+    );
 
     const coins = [
       {
@@ -267,7 +272,7 @@ export function getXdefiMethods(provider: BrowserProvider) {
             from,
             to,
             data: data || "0x",
-          } as Todo,
+          } as any,
         ]);
       }
       const contract = createContract(contractAddress, abi, contractProvider);
@@ -299,7 +304,7 @@ export function getXdefiMethods(provider: BrowserProvider) {
           from,
           to,
           data: data || "0x",
-        } as Todo,
+        } as any,
       ]);
     },
     sendTransaction: async (tx: EVMTxParams) => {
@@ -316,7 +321,7 @@ export function getXdefiMethods(provider: BrowserProvider) {
           from,
           to,
           data: data || "0x",
-        } as Todo,
+        } as any,
       ]);
     },
   };